Revert "Move specific operators closer to the application"
This reverts commit cc0afe6c9e8ac6fadf49295ec1b459e064090aa0.
