--- applications/endoscopy_tool_tracking/operators/lstm_tensor_rt_inference/python/CMakeLists.txt
@@ -1,23 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2022-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-include(pybind11_add_holohub_module)
-pybind11_add_holohub_module(
-    CPP_CMAKE_TARGET lstm_tensor_rt_inference
-    CLASS_NAME "LSTMTensorRTInferenceOp"
-    SOURCES lstm_tensor_rt_inference.cpp
-)
-
-target_include_directories(lstm_tensor_rt_inference_python PRIVATE ${Holohub_SOURCE_DIR}/operators)
