+++ gxf_extensions/CMakeLists.txt
@@ -18,4 +18,5 @@ set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib/gxf_extensions)
 
 add_holohub_extension(deltacast_videomaster)
 add_holohub_extension(emergent_source)
+add_holohub_extension(lstm_tensor_rt_inference)
 add_holohub_extension(yuan_qcap)
