+++ operators/lstm_tensor_rt_inference/metadata.json
